refactor: 패키지 이동 및 유틸 클래스 생성 제한
