+++ src/main/java/strings/util/StringValidator.java
@@ -1,6 +1,10 @@
-package strings;
+package strings.util;
 
 public final class StringValidator {
+
+    private StringValidator() {
+    }
+
     public static void validateString(final String str) {
         if (str == null || str.isBlank()) {
             throw new IllegalArgumentException("null이거나 공백인 문자열을 사용할 수 없습니다.");
