+++ src/test/java/strings/StringValidatorTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import strings.util.StringValidator;
 
 final class StringValidatorTest {
